feat(web): add sections directory structure and organize home components
