+++ apps/web/components/sections/home/journey-section.tsx
@@ -222,3 +222,4 @@ const ProjectJourney = () => {
 }
 
 export default ProjectJourney
+
