+++ apps/web/components/sections/home/participate-section.tsx
@@ -175,3 +175,4 @@ export const WhyInvestSection = () => {
 		</section>
 	)
 }
+
